add command options and fix codes based on second commnts.
